rename component_fields.py to field_widget.py
